Bug in badging

+++ ELSA/ConfigurationELSA.py
@@ -2772,7 +2772,7 @@ class AllSensors(AllObjects):
                                         currSensor.fields['acronym'] + u" = " + unicode(value))
                                 if key == 'G':
                                     try:
-                                        nfc_uid = hex(int(value)).zfill(14)
+                                        nfc_uid = hex(long(value)).zfill(14)
                                         elem = self.config.AllBarcodes.barcode_to_item(nfc_uid, "N")
                                         if elem:
                                             type = elem.get_type()
